requirements.txt
streamlit==1.28.0
supabase==2.16.0
requests==2.31.0
python-dotenv==1.0.0
Pillow==10.4.0
plotly==5.17.0
numpy==1.26.4
pandas==2.2.0
openai==1.12.0  # For OpenRouter API compatibility
python-jose==3.3.0  # For JWT handling
psycopg2-binary==2.9.9  # For PostgreSQL support
black==24.2.0  # For code formatting
